requirements.txt
asgiref==3.6.0
cloudinary==1.32.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.18
django-crispy-forms==2.0
gunicorn==20.1.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
